Move instruction decoding out of disassembler
The disassembler is now mainly the command line tool, which deals
with interpreting user input, formatting output, etc.

This allows us to add decoding logic for new cpus (the S2) and the
disassembler can then dynamically load the correct decoding module.
